requirements.txt
oracledb==3.2.0
pandas==2.3.1
plotly==6.2.0
jupyterlab==4.4.5
python-dotenv==1.1.1
